refactored code pinned to Cargo.lock polkadot-sdk hash for docker image
