+++ testing/zombienet-sdk-tests/tests/rococo_westend/mod.rs
@@ -1,0 +1,113 @@
+// Copyright (C) Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+//! The local Rococo <> Westend bridge pair: shared constants, the per-runtime typed operations
+//! (generated by the macros in [`ops`]), bridge-relayer reward queries and the network/bootstrap
+//! environment ([`environment`]). The individual tests live in `asset_transfer` and `free_headers`.
+//!
+//! A second bridge pair (e.g. `kusama_polkadot`) would be added as a sibling module built on the
+//! same [`crate::common`] infrastructure.
+
+use subxt::{OnlineClient, PolkadotConfig};
+
+// `1u64 << 60` — the amount the local chain specs endow every well-known account with. The
+// finality/parachain relayers (`//Charlie` / `//Dave`) only submit free or mandatory headers, so
+// their balance must stay exactly at this value throughout the tests.
+pub const ENDOWMENT: u128 = 1u128 << 60;
+
+/// Message lane shared by both Asset Hubs (`0x00000002`).
+pub const LANE_ID: [u8; 4] = [0, 0, 0, 2];
+/// Target XCM version configured for every remote location.
+pub const XCM_VERSION: u32 = 5;
+
+// Bridged chain ids (`*b"bhwd"` / `*b"bhro"`), used as reward keys on the Bridge Hubs.
+pub const BRIDGED_CHAIN_ID_BHWD: [u8; 4] = *b"bhwd";
+pub const BRIDGED_CHAIN_ID_BHRO: [u8; 4] = *b"bhro";
+
+// `6408de77..` / `e143f238..` — genesis hashes used as `NetworkId::ByGenesis(..)`.
+pub const ROCOCO_GENESIS_HASH: [u8; 32] = [
+	100, 8, 222, 119, 55, 197, 156, 35, 136, 144, 83, 58, 242, 88, 150, 162, 194, 6, 8, 216, 179,
+	128, 187, 1, 2, 154, 203, 57, 39, 129, 6, 62,
+];
+pub const WESTEND_GENESIS_HASH: [u8; 32] = [
+	225, 67, 242, 56, 3, 172, 80, 232, 246, 248, 230, 38, 149, 209, 206, 158, 78, 29, 104, 170, 54,
+	193, 205, 44, 253, 21, 52, 2, 19, 243, 66, 62,
+];
+
+// Para ids.
+pub const ASSET_HUB_PARA_ID: u32 = 1000;
+pub const BRIDGE_HUB_ROCOCO_PARA_ID: u32 = 1013;
+pub const BRIDGE_HUB_WESTEND_PARA_ID: u32 = 1002;
+
+// Sovereign / reward accounts funded on the Bridge Hubs.
+const ASSET_HUB_SOVEREIGN_AT_BRIDGE_HUB: &str = "5Eg2fntNprdN3FgH4sfEaaZhYtddZQSQUqvYJ1f2mLtinVhV";
+const BHR_LANE_THIS_CHAIN: &str = "5EHnXaT5GApse1euZWj9hycMbgjKBCNQL9WEwScL8QDx6mhK";
+const BHR_LANE_BRIDGED_CHAIN: &str = "5EHnXaT5Tnt4A8aiP9CsuAFRhKPjKZJXRrj4a3mtihFvKpTi";
+const BHW_LANE_THIS_CHAIN: &str = "5EHnXaT5GApry9tS6yd1FVusPq8o8bQJGCKyvXTFCoEKk5Z9";
+const BHW_LANE_BRIDGED_CHAIN: &str = "5EHnXaT5Tnt3VGpEvc6jSgYwVToDGxLRMuYoZ8coo6GHyWbR";
+
+// The amount funded onto each sovereign/reward account.
+const SOVEREIGN_FUNDING: u128 = 100_000_000_000_000;
+
+#[macro_use]
+mod ops;
+
+relay_ops!(relay_rococo, rococo, rococo_runtime);
+relay_ops!(relay_westend, westend, westend_runtime);
+asset_hub_ops!(asset_hub_rococo, asset_hub_rococo);
+asset_hub_ops!(asset_hub_westend, asset_hub_westend);
+bridge_hub_ops!(bridge_hub_rococo, bridge_hub_rococo);
+bridge_hub_ops!(bridge_hub_westend, bridge_hub_westend);
+
+/// Reads `bridgeRelayers.relayerRewards(relayer, RewardsAccountParams)` on Bridge Hub Rococo.
+pub async fn bridge_hub_rococo_relayer_reward(
+	client: &OnlineClient<PolkadotConfig>,
+	relayer: subxt::utils::AccountId32,
+) -> Result<Option<u128>, anyhow::Error> {
+	use crate::bridge_hub_rococo::runtime_types::{
+		bp_messages::lane::LegacyLaneId,
+		bp_relayers::{RewardsAccountOwner, RewardsAccountParams},
+	};
+	let reward = RewardsAccountParams {
+		owner: RewardsAccountOwner::ThisChain,
+		bridged_chain_id: BRIDGED_CHAIN_ID_BHWD,
+		lane_id: LegacyLaneId(LANE_ID),
+	};
+	let addr = crate::bridge_hub_rococo::storage()
+		.bridge_relayers()
+		.relayer_rewards(relayer, reward);
+	Ok(client.storage().at_latest().await?.fetch(&addr).await?)
+}
+
+/// Reads `bridgeRelayers.relayerRewards(relayer, BridgeReward::RococoWestend(..))` on Bridge Hub
+/// Westend (the reward kind is wrapped in the runtime's `BridgeReward` enum there).
+pub async fn bridge_hub_westend_relayer_reward(
+	client: &OnlineClient<PolkadotConfig>,
+	relayer: subxt::utils::AccountId32,
+) -> Result<Option<u128>, anyhow::Error> {
+	use crate::bridge_hub_westend::runtime_types::{
+		bp_messages::lane::LegacyLaneId,
+		bp_relayers::{RewardsAccountOwner, RewardsAccountParams},
+		bridge_hub_westend_runtime::bridge_common_config::BridgeReward,
+	};
+	let reward = BridgeReward::RococoWestend(RewardsAccountParams {
+		owner: RewardsAccountOwner::ThisChain,
+		bridged_chain_id: BRIDGED_CHAIN_ID_BHRO,
+		lane_id: LegacyLaneId(LANE_ID),
+	});
+	let addr = crate::bridge_hub_westend::storage()
+		.bridge_relayers()
+		.relayer_rewards(relayer, reward);
+	Ok(client.storage().at_latest().await?.fetch(&addr).await?)
+}
+
+mod environment;
+pub use environment::BridgeTestEnv;
+
+// Re-exported so the per-pair tests can reach the generic helpers via `crate::rococo_westend::*`.
+pub use crate::common::utils::{
+	count_synced_headers, dev_account, dev_public, free_balance_at, retry_until,
+};
+
+mod asset_transfer;
+mod free_headers;
